+++ sdk/identity/identity/src/credentials/deviceCodeCredential.ts
@@ -1,47 +1,17 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import { AccessToken, TokenCredential, GetTokenOptions } from "@azure/core-http";
-import { AuthenticationRequired, MsalClient } from "../client/msalClient";
-import { createSpan } from "../util/tracing";
-import { credentialLogger, formatError, formatSuccess } from "../util/logging";
-import { AuthenticationErrorName } from "../client/errors";
-import { CanonicalCode } from "@opentelemetry/api";
-import { TokenCredentialOptions } from "../client/identityClient";
 
-import { DeviceCodeRequest } from "@azure/msal-node";
-import { checkTenantId } from "../util/checkTenantId";
-import { DeveloperSignOnClientId } from "../constants";
-
-/**
- * Provides the user code and verification URI where the code must be
- * entered.  Also provides a message to display to the user which
- * contains an instruction with these details.
- */
-export interface DeviceCodeInfo {
-  /**
-   * The device code that the user must enter into the verification page.
-   */
-  userCode: string;
-
-  /**
-   * The verification URI to which the user must navigate to enter the device
-   * code.
-   */
-  verificationUri: string;
-
-  /**
-   * A message that may be shown to the user to instruct them on how to enter
-   * the device code in the page specified by the verification URI.
-   */
-  message: string;
-}
-
-/**
- * Defines the signature of a callback which will be passed to
- * DeviceCodeCredential for the purpose of displaying authentication
- * details to the user.
- */
-export type DeviceCodePromptCallback = (deviceCodeInfo: DeviceCodeInfo) => void;
+import { AccessToken, GetTokenOptions, TokenCredential } from "@azure/core-http";
+import { credentialLogger } from "../util/logging";
+import { MsalDeviceCode } from "../msal/nodeFlows/msalDeviceCode";
+import { MsalFlow } from "../msal/flows";
+import { AuthenticationRecord } from "../msal/types";
+import { trace } from "../util/tracing";
+import {
+  DeviceCodeCredentialOptions,
+  DeviceCodeInfo,
+  DeviceCodePromptCallback
+} from "./deviceCodeCredentialOptions";
 
 const logger = credentialLogger("DeviceCodeCredential");
 
@@ -58,8 +28,8 @@ export function defaultDeviceCodePromptCallback(deviceCodeInfo: DeviceCodeInfo):
  * that the user can enter into https://microsoft.com/devicelogin.
  */
 export class DeviceCodeCredential implements TokenCredential {
-  private userPromptCallback: DeviceCodePromptCallback;
-  private msalClient: MsalClient;
+  private msalFlow: MsalFlow;
+  private disableAutomaticAuthentication?: boolean;
 
   /**
    * Creates an instance of DeviceCodeCredential with the details needed
@@ -68,39 +38,42 @@ export class DeviceCodeCredential implements TokenCredential {
    * @param tenantId - The Azure Active Directory tenant (directory) ID or name.
    *                   The default value is 'organizations'.
    *                   'organizations' may be used when dealing with multi-tenant scenarios.
+   *                   Users can also pass the options as the first parameter, and skip the other parammeters entirely.
    * @param clientId - The client (application) ID of an App Registration in the tenant.
    *                   By default we will try to use the Azure CLI's client ID to authenticate.
    * @param userPromptCallback - A callback function that will be invoked to show
-                               {@link DeviceCodeInfo} to the user. If left unassigned, we will automatically log the device code information and the authentication instructions in the console.
-   * @param options - Options for configuring the client which makes the authentication request.
+   *                             {@link DeviceCodeInfo} to the user. If left unassigned, we will automatically log the device code information and the authentication instructions in the console.
+   *                             Users can also pass the options as the third parameter, and skip the prompt callback entirely.
+   * @param options - Options for configuring the client which makes the authentication requests.
    */
+  constructor(options?: DeviceCodeCredentialOptions);
   constructor(
-    tenantId: string = "organizations",
-    clientId: string = DeveloperSignOnClientId,
-    userPromptCallback: DeviceCodePromptCallback = defaultDeviceCodePromptCallback,
-    options?: TokenCredentialOptions
+    tenantId?: string,
+    clientId?: string,
+    userPromptCallback?: DeviceCodePromptCallback,
+    options?: DeviceCodeCredentialOptions
+  );
+  constructor(
+    tenantIdOrOptions?: string | DeviceCodeCredentialOptions,
+    clientId?: string,
+    userPromptCallback?: DeviceCodePromptCallback,
+    options?: DeviceCodeCredentialOptions
   ) {
-    checkTenantId(logger, tenantId);
-
-    this.userPromptCallback = userPromptCallback;
-
-    let authorityHost;
-    if (options && options.authorityHost) {
-      if (options.authorityHost.endsWith("/")) {
-        authorityHost = options.authorityHost + tenantId;
-      } else {
-        authorityHost = options.authorityHost + "/" + tenantId;
-      }
+    let tenantId: string | undefined;
+    if (typeof tenantIdOrOptions === "string") {
+      tenantId = tenantIdOrOptions;
     } else {
-      authorityHost = "https://login.microsoftonline.com/" + tenantId;
+      options = tenantIdOrOptions;
     }
-
-    this.msalClient = new MsalClient(
-      { clientId: clientId, authority: authorityHost },
-      false,
-      undefined,
-      options
-    );
+    this.msalFlow = new MsalDeviceCode({
+      ...options,
+      logger,
+      clientId,
+      tenantId,
+      userPromptCallback: userPromptCallback || defaultDeviceCodePromptCallback,
+      tokenCredentialOptions: options || {}
+    });
+    this.disableAutomaticAuthentication = options?.disableAutomaticAuthentication;
   }
 
   /**
@@ -109,69 +82,44 @@ export class DeviceCodeCredential implements TokenCredential {
    * return null.  If an error occurs during authentication, an {@link AuthenticationError}
    * containing failure details will be thrown.
    *
+   * If the user provided the option `disableAutomaticAuthentication`,
+   * once the token can't be retrieved silently,
+   * this method won't attempt to request user interaction to retrieve the token.
+   *
    * @param scopes - The list of scopes for which the token will have access.
    * @param options - The options used to configure any requests this
    *                TokenCredential implementation might make.
    */
-  async getToken(
-    scopes: string | string[],
-    options?: GetTokenOptions
-  ): Promise<AccessToken | null> {
-    const { span } = createSpan("DeviceCodeCredential-getToken", options);
-
-    const scopeArray = typeof scopes === "object" ? scopes : [scopes];
-
-    const deviceCodeRequest = {
-      deviceCodeCallback: this.userPromptCallback,
-      scopes: scopeArray
-    };
-
-    logger.info(`DeviceCodeCredential invoked. Scopes: ${scopeArray.join(", ")}`);
-
-    return this.msalClient.acquireTokenFromCache(scopeArray).catch(async (e) => {
-      if (e instanceof AuthenticationRequired) {
-        try {
-          const token = await this.acquireTokenByDeviceCode(deviceCodeRequest, scopeArray);
-          logger.getToken.info(formatSuccess(scopeArray));
-          return token;
-        } catch (err) {
-          const code =
-            err.name === AuthenticationErrorName
-              ? CanonicalCode.UNAUTHENTICATED
-              : CanonicalCode.UNKNOWN;
-          span.setStatus({
-            code,
-            message: err.message
-          });
-          logger.getToken.info(formatError(scopeArray, err));
-          throw err;
-        } finally {
-          span.end();
-        }
-      } else {
-        throw e;
-      }
+  async getToken(scopes: string | string[], options: GetTokenOptions = {}): Promise<AccessToken> {
+    return trace(`${this.constructor.name}.getToken`, options, async (newOptions) => {
+      const arrayScopes = typeof scopes === "object" ? scopes : [scopes];
+      return this.msalFlow.getToken(arrayScopes, {
+        ...newOptions,
+        disableAutomaticAuthentication: this.disableAutomaticAuthentication
+      });
     });
   }
 
-  private async acquireTokenByDeviceCode(
-    deviceCodeRequest: DeviceCodeRequest,
-    scopes: string[]
-  ): Promise<AccessToken | null> {
-    try {
-      const deviceResponse = await this.msalClient.acquireTokenByDeviceCode(deviceCodeRequest);
-      if (deviceResponse && deviceResponse.expiresOn) {
-        const expiresOnTimestamp = deviceResponse.expiresOn.getTime();
-        logger.getToken.info(formatSuccess(scopes));
-        return {
-          expiresOnTimestamp,
-          token: deviceResponse.accessToken
-        };
-      } else {
-        throw new Error("Did not receive token with a valid expiration");
-      }
-    } catch (error) {
-      throw new Error(`Device Authentication Error "${JSON.stringify(error)}"`);
-    }
+  /**
+   * Authenticates with Azure Active Directory and returns an access token if
+   * successful.  If authentication cannot be performed at this time, this method may
+   * return null.  If an error occurs during authentication, an {@link AuthenticationError}
+   * containing failure details will be thrown.
+   *
+   * If the token can't be retrieved silently, this method will require user interaction to retrieve the token.
+   *
+   * @param scopes - The list of scopes for which the token will have access.
+   * @param options - The options used to configure any requests this
+   *                  TokenCredential implementation might make.
+   */
+  async authenticate(
+    scopes: string | string[],
+    options: GetTokenOptions = {}
+  ): Promise<AuthenticationRecord | undefined> {
+    return trace(`${this.constructor.name}.authenticate`, options, async (newOptions) => {
+      const arrayScopes = typeof scopes === "object" ? scopes : [scopes];
+      await this.msalFlow.getToken(arrayScopes, newOptions);
+      return this.msalFlow.getActiveAccount();
+    });
   }
 }