+++ sdk/identity/identity/src/credentials/usernamePasswordCredential.ts
@@ -1,15 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import qs from "qs";
-import { TokenCredential, GetTokenOptions, AccessToken } from "@azure/core-http";
-import { TokenCredentialOptions, IdentityClient } from "../client/identityClient";
-import { createSpan } from "../util/tracing";
-import { AuthenticationErrorName } from "../client/errors";
-import { CanonicalCode } from "@opentelemetry/api";
-import { credentialLogger, formatSuccess, formatError } from "../util/logging";
-import { getIdentityTokenEndpointSuffix } from "../util/identityTokenEndpoint";
-import { checkTenantId } from "../util/checkTenantId";
+import { AccessToken, GetTokenOptions, TokenCredential } from "@azure/core-http";
+import { credentialLogger } from "../util/logging";
+import { MsalUsernamePassword } from "../msal/nodeFlows/msalUsernamePassword";
+import { MsalFlow } from "../msal/flows";
+import { trace } from "../util/tracing";
+import { AuthenticationRecord } from "../msal/types";
+import { UsernamePasswordCredentialOptions } from "./usernamePasswordCredentialOptions";
 
 const logger = credentialLogger("UsernamePasswordCredential");
 
@@ -19,38 +17,40 @@ const logger = credentialLogger("UsernamePasswordCredential");
  * trust so you should only use it when other, more secure credential
  * types can't be used.
  */
+// We'll be using InteractiveCredential as the base of this class, which requires us to support authenticate(),
+// to reduce the number of times we send the password over the network.
 export class UsernamePasswordCredential implements TokenCredential {
-  private identityClient: IdentityClient;
-  private tenantId: string;
-  private clientId: string;
-  private username: string;
-  private password: string;
+  private msalFlow: MsalFlow;
+  private disableAutomaticAuthentication?: boolean;
 
   /**
    * Creates an instance of the UsernamePasswordCredential with the details
    * needed to authenticate against Azure Active Directory with a username
    * and password.
    *
-   * @param tenantIdOrName - The Azure Active Directory tenant (directory) ID or name.
+   * @param tenantId - The Azure Active Directory tenant (directory).
    * @param clientId - The client (application) ID of an App Registration in the tenant.
    * @param username - The user account's e-mail address (user name).
    * @param password - The user account's account password
    * @param options - Options for configuring the client which makes the authentication request.
    */
   constructor(
-    tenantIdOrName: string,
+    tenantId: string,
     clientId: string,
     username: string,
     password: string,
-    options?: TokenCredentialOptions
+    options: UsernamePasswordCredentialOptions = {}
   ) {
-    checkTenantId(logger, tenantIdOrName);
-
-    this.identityClient = new IdentityClient(options);
-    this.tenantId = tenantIdOrName;
-    this.clientId = clientId;
-    this.username = username;
-    this.password = password;
+    this.msalFlow = new MsalUsernamePassword({
+      ...options,
+      logger,
+      clientId,
+      tenantId,
+      username,
+      password,
+      tokenCredentialOptions: options || {}
+    });
+    this.disableAutomaticAuthentication = options?.disableAutomaticAuthentication;
   }
 
   /**
@@ -59,54 +59,44 @@ export class UsernamePasswordCredential implements TokenCredential {
    * return null.  If an error occurs during authentication, an {@link AuthenticationError}
    * containing failure details will be thrown.
    *
+   * If the user provided the option `disableAutomaticAuthentication`,
+   * once the token can't be retrieved silently,
+   * this method won't attempt to request user interaction to retrieve the token.
+   *
    * @param scopes - The list of scopes for which the token will have access.
    * @param options - The options used to configure any requests this
    *                TokenCredential implementation might make.
    */
-  public async getToken(
-    scopes: string | string[],
-    options?: GetTokenOptions
-  ): Promise<AccessToken | null> {
-    const { span, updatedOptions } = createSpan("UsernamePasswordCredential-getToken", options);
-    try {
-      const urlSuffix = getIdentityTokenEndpointSuffix(this.tenantId);
-      const webResource = this.identityClient.createWebResource({
-        url: `${this.identityClient.authorityHost}/${this.tenantId}/${urlSuffix}`,
-        method: "POST",
-        disableJsonStringifyOnBody: true,
-        deserializationMapper: undefined,
-        body: qs.stringify({
-          response_type: "token",
-          grant_type: "password",
-          client_id: this.clientId,
-          username: this.username,
-          password: this.password,
-          scope: typeof scopes === "string" ? scopes : scopes.join(" ")
-        }),
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/x-www-form-urlencoded"
-        },
-        abortSignal: options && options.abortSignal,
-        spanOptions: updatedOptions?.tracingOptions?.spanOptions
+  async getToken(scopes: string | string[], options: GetTokenOptions = {}): Promise<AccessToken> {
+    return trace(`${this.constructor.name}.getToken`, options, async (newOptions) => {
+      const arrayScopes = typeof scopes === "object" ? scopes : [scopes];
+      return this.msalFlow.getToken(arrayScopes, {
+        ...newOptions,
+        disableAutomaticAuthentication: this.disableAutomaticAuthentication
       });
+    });
+  }
 
-      const tokenResponse = await this.identityClient.sendTokenRequest(webResource);
-      logger.getToken.info(formatSuccess(scopes));
-      return (tokenResponse && tokenResponse.accessToken) || null;
-    } catch (err) {
-      const code =
-        err.name === AuthenticationErrorName
-          ? CanonicalCode.UNAUTHENTICATED
-          : CanonicalCode.UNKNOWN;
-      span.setStatus({
-        code,
-        message: err.message
-      });
-      logger.getToken.info(formatError(scopes, err));
-      throw err;
-    } finally {
-      span.end();
-    }
+  /**
+   * Authenticates with Azure Active Directory and returns an access token if
+   * successful.  If authentication cannot be performed at this time, this method may
+   * return null.  If an error occurs during authentication, an {@link AuthenticationError}
+   * containing failure details will be thrown.
+   *
+   * If the token can't be retrieved silently, this method will require user interaction to retrieve the token.
+   *
+   * @param scopes - The list of scopes for which the token will have access.
+   * @param options - The options used to configure any requests this
+   *                  TokenCredential implementation might make.
+   */
+  async authenticate(
+    scopes: string | string[],
+    options: GetTokenOptions = {}
+  ): Promise<AuthenticationRecord | undefined> {
+    return trace(`${this.constructor.name}.authenticate`, options, async (newOptions) => {
+      const arrayScopes = typeof scopes === "object" ? scopes : [scopes];
+      await this.msalFlow.getToken(arrayScopes, newOptions);
+      return this.msalFlow.getActiveAccount();
+    });
   }
 }